Extract functions into their own files
