+++ src/support_code_library_builder/get_definition_line_and_uri.ts
@@ -1,10 +1,8 @@
-import { ParameterType } from '@cucumber/cucumber-expressions'
 import path from 'path'
 import StackTrace from 'stacktrace-js'
 import { isFileNameInCucumber } from '../stack_trace_filter'
 import { doesHaveValue, valueOrDefault } from '../value_checker'
 import { ILineAndUri } from '../types'
-import { IParameterTypeDefinition } from './types'
 
 export function getDefinitionLineAndUri(cwd: string): ILineAndUri {
   let line: number
@@ -28,23 +26,4 @@ export function getDefinitionLineAndUri(cwd: string): ILineAndUri {
     line: valueOrDefault(line, 0),
     uri: valueOrDefault(uri, 'unknown'),
   }
-}
-
-export function buildParameterType({
-  name,
-  regexp,
-  transformer,
-  useForSnippets,
-  preferForRegexpMatch,
-}: IParameterTypeDefinition<any>): ParameterType<any> {
-  if (typeof useForSnippets !== 'boolean') useForSnippets = true
-  if (typeof preferForRegexpMatch !== 'boolean') preferForRegexpMatch = false
-  return new ParameterType(
-    name,
-    regexp,
-    null,
-    transformer,
-    useForSnippets,
-    preferForRegexpMatch
-  )
 }
